@@ -1,49 +1,38 @@
 {
-  "name": "runawfe-professional",
-  "version": "1.0.0",
-  "description": "",
+  "name": "runawfe",
+  "version": "0.0.1",
   "private": true,
+  "type": "module",
   "scripts": {
-    "debug": "webpack serve --config webpack.dev.js",
-    "build": "webpack --config webpack.prod.js",
-    "start": "http-server dist"
+    "dev": "vite --host 127.0.0.1",
+    "build": "vite build --base /spa",
+    "preview": "vite preview",
+    "type-check": "vue-tsc --build --force",
+    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --ignore-path .gitignore"
+  },
+  "dependencies": {
+    "pinia": "^2.1.7",
+    "vue": "^3.3.11",
+    "vue-router": "^4.2.5",
+    "vuetify": "^3.5.1"
   },
   "devDependencies": {
-    "@mdi/font": "^7.1.96",
-    "@types/lodash": "^4.14.168",
-    "@types/webpack-env": "^1.16.0",
-    "axios": "^0.21.1",
-    "clean-webpack-plugin": "^3.0.0",
-    "core-js": "^3.9.1",
-    "css-loader": "^6.7.3",
-    "deepmerge": "^4.2.2",
-    "html-webpack-plugin": "^4.5.2",
-    "http-server": "^0.12.3",
-    "lodash": "^4.17.21",
-    "mini-css-extract-plugin": "^2.7.2",
-    "node-sass": "^5.0.0",
-    "regenerator-runtime": "^0.13.7",
-    "sass": "^1.32.8",
-    "sass-loader": "^10.1.1",
-    "style-loader": "^2.0.0",
+    "@mdi/font": "^7.4.47",
+    "@rushstack/eslint-patch": "^1.3.3",
+    "@tsconfig/node18": "^18.2.2",
+    "@types/lodash": "^4.14.202",
+    "@types/node": "^18.19.10",
+    "@vitejs/plugin-vue": "^4.5.2",
+    "@vue/eslint-config-typescript": "^12.0.0",
+    "@vue/tsconfig": "^0.5.0",
+    "eslint": "^8.49.0",
+    "eslint-plugin-vue": "^9.17.0",
+    "js-base64": "^3.7.6",
+    "npm-run-all2": "^6.1.1",
+    "sass": "^1.70.0",
     "swagger-client": "^3.13.1",
-    "ts-loader": "^8.0.18",
-    "typescript": "^4.2.3",
-    "vue": "^2.6.12",
-    "vue-loader": "^15.9.6",
-    "vue-meta": "^2.4.0",
-    "vue-property-decorator": "^9.1.2",
-    "vue-router": "^3.5.1",
-    "vue-style-loader": "^4.1.3",
-    "vue-template-compiler": "^2.6.12",
-    "vuetify": "^2.4.6",
-    "vuetify-loader": "^1.7.2",
-    "vuex": "^3.6.2",
-    "vuex-pathify": "^1.4.5",
-    "vuex-router-sync": "^5.0.0",
-    "webpack": "^5.24.4",
-    "webpack-cli": "^4.5.0",
-    "webpack-dev-server": "^3.11.2",
-    "webpack-merge": "^5.7.3"
+    "typescript": "~5.3.0",
+    "vite": "^5.0.10",
+    "vue-tsc": "^1.8.25"
   }
 }
