@@ -35,9 +35,9 @@
 </template>
 
 <script lang="ts">
-import Vue from 'vue';
+import { defineComponent } from 'vue';
 
-export default Vue.extend({
+export default defineComponent({
   name: "ProfileUserCard",
   props: {
     user: Object,
